+++ packages/core/src/node/FipsNode.ts
@@ -179,6 +179,7 @@ export class FipsNode {
       throw err;
     }
     this.started = true;
+    this.sessionManager.start();
     this.heartbeatTimer = setInterval(() => {
       void this.sendHeartbeats();
     }, this.heartbeatIntervalMs);